Add a signal module (starting with an implementation of the Kaiser function)
This module will have signal processing related functions (e.g. to generate windows, resampling, etc).

The first set of functions added to this module allow you to calculate the Kaiser window.
To do so, we also add an implementation of the Bessel function of the first kind of order 0 (i0).

The code in this commit is based on the code for the corresponding functions in the scipy library.